removed skeleton-files directory
