+++ .gitignore
@@ -1,1 +1,2 @@
 talbot
+update_local.sh
