--- skeleton-files/docker-compose.yaml
@@ -1,8 +1,0 @@
-version: "3.9"
-
-services:
-  server:
-    image: server
-    build: .
-    ports:
-      - "4000:4000"
